THIS IS THE INITIAL COMMIT FOR THE CYRILLIC-KEYBOARD BRANCH
package.json was updated and package-lock.json was created due to the
installation of the following packages:
    node-sass
    bootstrap
    react-bootstrap

I created a non-functional CyrillicKeyboard component; it is
non-functional because it currently does not raise the click event.

The keyboard component uses the new _variables.scss and
cyrillic-keyboard-component.scss files. It has basically styling as in
the GA Russian Flashcards project. It is using React Bootstrap
buttons.

App.js was updated to import and use the new keyboard, and to
overwrite the create-react-app version of App.js.

- new file:   package-lock.json
- modified:   package.json
- modified:   src/App.js
- new file:   src/components/cyrillic-keyboard/_variables.scss
- new file:   src/components/cyrillic-keyboard/
              cyrillic-keyboard-component.jsx
- new file:   src/components/cyrillic-keyboard/
              cyrillic-keyboard-component.scss