I REFACTORED THE CYRILLIC KEYBOARD
It bothered me that I exposed an event to the parent componet. So I
refactored the callback in App.js to pass a callback that takes a
single parameter which represents the value of the Cyrillic key that
was pressed. The keyboard now does the work of extracting the value
from the synthetic event.

This now works as I intend. I shall revisit to improve the styling
of the keyboard. I removed the border radius for now since it seems
to big. I also removed the inline-block style for the entier keyboard.
It should be parent components that determine how to display the
keyboard, not the keyboard itself.