BASIC REACT ROUTER ROUTING IS WORKING
Basic routing is working by manually typing the routes in the URL.

I added the react-router-dom package.

index.js is now using the HashRouter.

App.jsx was updated to host all of the routes for the app. I commented
the code for the input control and the Cyrillic Keyboard. That was
there so that I could remember what I did when I move that code into
their view components.

I added do-nothing "view" components that represent each view in the
app. They just render a heading to that I could manually test the
routing by typing the URL in the browser. Nothing yet allows navigation
between the views with respect to links or buttons.