ROUTING IS ALMOST COMPLETE
I fixed the route in App.js for the home view. It now uses "exact" in
the route. The result is that I no longer need to navigate to "/home".
Now "/" just works.

All of the view components which I created in my previous commit now
have buttons that support routing as in the original app. I can
navigate to and from views at will.

I might need to add to routing when I support signin / signout etc.