THE INITIAL COMMIT FOR THE CREATE FLASHCARD VIEW
This is the initial commit for the create flashcard view. This will
be the exemplar for the other views.

Code that was commented in App.js was removed and put into the
create flashcard view. It is now possible to type into the keyboard
and see Russian text in the russian textbox. I'll work on styling, and
I must determine where to do data access.