THE CYRILLIC KEYBOARD IS NOW ABLE TO BE DISABLED
I added support to the Cyrillic keyboard for it being disabled.
There are now two props; disabled and keyboardPressHandler.
When disabled is true, the keyboard is disabled and you cannot type
on it. When false, you can use the keyboard to type Russian.

keyboardPressHandler is a callback function from a parent component
that handles the Cyrilli key presses. It simply translates the keypress
events to a Cyrillic character. This has been supported since the
initial commit of the keyboard.