THE STATUS MESSAGE AREA COMPONENT IS WORKING.
This component is effectively working. It may require a tweak or two.

App.scss was updated with the styling for the status message area as it
was in the GA project @ version.

App.jsx uses the render method in the route in order to pass a message
to the message component.

StatusMessages now takes a message to be displayed. It is also now a
paragraph.