THE HOME VIEW IS NEARLY COMPLETE
I modified index.html to change the title element to "Russian
Flashcards".

I styled the home page using React Bootstrap grid.

I added the images folder under source, and added the image of
St. Basil's Cathedral. Create React App requires that all relative
paths must be within "src". So that is why I put the folder and image
there:
https://create-react-app.dev/docs/adding-images-fonts-and-files

Of course home-view-component.jsx is was altered. And so was App.jsx.
App.jsx is using the render method in the route definition for the
home screen. In the render method is passes a callback for a method that
sets state for the StatusMessages component. This will cause a render
to happen when the message changes.

For HomeView, I had to add a history parameter to HomeView in order
for navigation to work inside of HomeView. It also takes the callback
for setting a message in StatusMessages.

The result is that status messages can be written from the home view,
and navigation is working to signup and signin views.

I think that this is mostly complete, certainly it is functionally.