THE SIGNIN VIEW IS MOSTLY WORKING
What is missing is the web service call to acutally sigin and
obtain a token.

App.jsx was updated to use a rendered route for Signin view in order
to pass history and the callback for the status message component so
that sigin view can display messages to the app.

The sigin button in SigninView component for now enables only when the
email and password fields are not empty. The input fields are
controlled by React and their values can be set by typing and extracted
when the web service calls will be implemented.