I FIXED THE VIEW FLASHCARDS VIEW
Based upon this morning's lesson at GA, I learned that in React,
form controls are "controlled".

I was using references, but that is not needed. The English keyboard
now has a value in state. The Form input uses that, and I have a
callback to handle changes in the English input field. Everything works
and it follows the pattern for doing what I am doing.