LOGGING INTO AND OUT OF THE APP NOW WORKS
I copied the store.js file from the original project but modified it
for ES6 modules. This is used to store the token.

I also copied the apiConfig file from the GA project that gives the
URLs to the backend.

I installed the Axios npm package.

I created a “models” folder and added data-model.js. This hides the
details of data access from the React components, and that Axios is
being used under the covers.

SigninView now support signing into the app.

OptionsView now support logging out of the app. Login and logout
works.

I added the .env file to change the port number from 3000 to 7165.
CORS is setup on the backend to only allow traffic from 7165 from
the GA project #2. I struggled with this but solved it.