UPDATE FLASHCARDS IS FEATURE COMPLETE
The component is feature complete. This blows away what I did for
project 2 for GA. You had to know a flashcard and search for it. Now
all of the flashcards are displayed. You click one it autofills the
input fields. The buttons do the right things, and the keyboard
disables when the English input has focus.

The data-model.js file now has the update service.

I had to add styling to flashcard-views.scss to style the flashcards.