VIEW FLASHCARDS IS NOW COMPLETE
First, I fixed the Russian keyboard. You could not fix mistakes with the
Russian input field because it is controlled by React. So I replaced
the enter key with a backspace key. Now you can backspace in Russian
input fields to fix mistakes in view.

ViewFlashcards component is using the updated Cyrillic keyboard. I
had to add a handler in ViewFlashcards to handle the backspace. In a
future version, components using the keyboard should pass a string to
the keyboard and let the keyboard component handle the string.

I had to tweak the styling for the backspace key since the key was not
wide enough.