CREATE FLASHCARDS IS FEATURE COMPLETE
The CreateFlashcard component was updated to support the following:

The new backspace text based on Arsyni’s feedback. This component now
permits backspace to work with the Cyrillic keyboard.

The handleEmptyCyrillicOnChange method for the russian input field.
This exists because the input field is bound to the soft keyboard, and
you will get an error from React. So this is a do-nothing method to
satisfy React. Without it, the field becomes read-only.

The Cyrillic keyboard now enables / disables at the appopriate times.
You cannot type English in the Russian input control, and the converse
is also true.