UPDATE FLASHCARD IS NOW FEATURE COMPLETE
I had to update styling in App.scss and flashcards-views.scss in order
to reduce the height of UpdateFlashcars to eliminate the vertical
scrollbar.

The cyrillic keyboard enables and disables as appropriate. I noticed a
small issue; you can create a flashcard inside of updateflashcard. I'll
fix it in a future release.